+++ CHANGES.rst
@@ -9,6 +9,10 @@ Changelog of lizard-auth-client
 
 - Sorting users on ``/sso/user_overview/`` alphabetically by email.
 
+- Added the function ``get_user_org_role_dict()`` to the models module.
+  The dictionary it returns will be used as JWT payload for
+  3Di services to share user permissions (roles).
+
 
 2.6 (2016-11-21)
 ----------------